Commit3-Sécurité_photo_final

+++ src/main/java/com/tchindaClovis/gestiondestock/services/impl/VenteServiceImpl.java
@@ -1,17 +1,15 @@
 package com.tchindaClovis.gestiondestock.services.impl;
 
-import com.tchindaClovis.gestiondestock.dto.CommandeClientDto;
-import com.tchindaClovis.gestiondestock.dto.LigneVenteDto;
-import com.tchindaClovis.gestiondestock.dto.VenteDto;
+import com.tchindaClovis.gestiondestock.dto.*;
 import com.tchindaClovis.gestiondestock.exception.EntityNotFoundException;
 import com.tchindaClovis.gestiondestock.exception.ErrorCodes;
 import com.tchindaClovis.gestiondestock.exception.InvalidEntityException;
-import com.tchindaClovis.gestiondestock.model.Article;
-import com.tchindaClovis.gestiondestock.model.LigneVente;
-import com.tchindaClovis.gestiondestock.model.Vente;
+import com.tchindaClovis.gestiondestock.exception.InvalidOperationException;
+import com.tchindaClovis.gestiondestock.model.*;
 import com.tchindaClovis.gestiondestock.repository.ArticleRepository;
 import com.tchindaClovis.gestiondestock.repository.LigneVenteRepository;
 import com.tchindaClovis.gestiondestock.repository.VenteRepository;
+import com.tchindaClovis.gestiondestock.services.MvtStockService;
 import com.tchindaClovis.gestiondestock.services.VenteService;
 import com.tchindaClovis.gestiondestock.validator.VenteValidator;
 import lombok.extern.slf4j.Slf4j;
@@ -19,6 +17,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -31,47 +30,51 @@ public class VenteServiceImpl implements VenteService {
     private VenteRepository venteRepository;
     private LigneVenteRepository ligneVenteRepository;
 
+    private MvtStockService mvtStockService;
+
     @Autowired
     public VenteServiceImpl(ArticleRepository articleRepository, VenteRepository
-            venteRepository, LigneVenteRepository ligneVenteRepository) {
+            venteRepository, LigneVenteRepository ligneVenteRepository, MvtStockService mvtStockService) {
         this.articleRepository = articleRepository;
         this.venteRepository = venteRepository;
         this.ligneVenteRepository = ligneVenteRepository;
+        this.mvtStockService = mvtStockService;
     }
 
     @Override
     public VenteDto save(VenteDto dto) {
         List<String> errors = VenteValidator.validate(dto);
-        if(!errors.isEmpty()){
-            log.error("Vente n'est pas valide");
-            throw new InvalidEntityException("Vente non valide", ErrorCodes.VENTE_NOT_VALID, errors);
+        if (!errors.isEmpty()) {
+            log.error("Ventes n'est pas valide");
+            throw new InvalidEntityException("L'objet vente n'est pas valide", ErrorCodes.VENTE_NOT_VALID, errors);
         }
 
         List<String> articleErrors = new ArrayList<>();
 
         dto.getLigneVentes().forEach(ligneVenteDto -> {
             Optional<Article> article = articleRepository.findById(ligneVenteDto.getArticle().getId());
-            if(article.isEmpty()){
-                articleErrors.add("Aucun article avec l'ID" + ligneVenteDto.getArticle().getId() +
-                        "n'a été trouvé dans la BDD");
+            if (article.isEmpty()) {
+                articleErrors.add("Aucun article avec l'ID " + ligneVenteDto.getArticle().getId() + " n'a ete trouve dans la BDD");
             }
         });
 
-        if(!articleErrors.isEmpty()){
-            log.error("one or more articles were not found in the DB, {}", errors);
-            throw new InvalidEntityException("Un ou plusieurs articles n'ont pas été trouvé dan la BDD", ErrorCodes.VENTE_NOT_VALID, errors);
+        if (!articleErrors.isEmpty()) {
+            log.error("One or more articles were not found in the DB, {}", errors);
+            throw new InvalidEntityException("Un ou plusieurs articles n'ont pas ete trouve dans la BDD", ErrorCodes.VENTE_NOT_VALID, errors);
         }
 
-        Vente savedVente = venteRepository.save(VenteDto.toEntity(dto));
+        Vente savedVentes = venteRepository.save(VenteDto.toEntity(dto));
 
         dto.getLigneVentes().forEach(ligneVenteDto -> {
             LigneVente ligneVente = LigneVenteDto.toEntity(ligneVenteDto);
-            ligneVente.setVente(savedVente);
+            ligneVente.setVente(savedVentes);
             ligneVenteRepository.save(ligneVente);
+            updateMvtStock(ligneVente);
         });
 
-        return VenteDto.fromEntity(savedVente);
+        return VenteDto.fromEntity(savedVentes);
     }
+
 
     @Override
     public VenteDto findById(Integer id) {
@@ -109,10 +112,27 @@ public class VenteServiceImpl implements VenteService {
 
     @Override
     public void delete(Integer id) {
-        if(id == null){
+        if (id == null) {
             log.error("Vente ID is NULL");
-            return ;
+            return;
+        }
+        List<LigneVente> ligneVentes = ligneVenteRepository.findAllByVenteId(id);
+        if (!ligneVentes.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une vente ...",
+                    ErrorCodes.VENTE_ALREADY_IN_USE);
         }
         venteRepository.deleteById(id);
+    }
+
+    private void updateMvtStock(LigneVente lig) {
+        MvtStockDto mvtStockDto = MvtStockDto.builder()
+                .article(ArticleDto.fromEntity(lig.getArticle()))
+                .dateMvt(Instant.now())
+                .typeMvt(ETypeMvtStock.SORTIE)
+                .sourceMvt(ESourceMvtStock.VENTE)
+                .quantite(lig.getQuantite())
+                .idEntreprise(lig.getIdEntreprise())
+                .build();
+        mvtStockService.sortieStock(mvtStockDto);
     }
 }
